% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exac.R
\name{exac}
\alias{exac}
\title{exac}
\usage{
exac(var)
}
\arguments{
\item{var}{A single variant in the format chr-pos-ref-alt}
}
\value{
A vector of the four items to look up
}
\description{
Query ExAC database for Consequence, Allele Frequency, PolyPhen Score and
gene SYMBOL
}
\author{
Alexander Hopkins
}
