% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_vcf.R
\name{simplify_vcf}
\alias{simplify_vcf}
\title{simplify_vcf}
\usage{
simplify_vcf(vcf)
}
\arguments{
\item{vcf}{An ExpandedVCF object}
}
\value{
An ExpandedVCF object of the same length as the input with necessary
records simplified
}
\description{
simplify_vcf identifies records which may have been expanded when converting
CollapsedVCF to ExpandedVCF, simplifies those records, and returns a
corrected vcf of the same length as the input
}
\seealso{
\code{\link{simplify_variant}}
}
\author{
Alexander Hopkins
}
